inline ERC721Full behavior
